rollback method move
