+++ generators/base-application/support/index.ts
@@ -22,6 +22,7 @@ export * from './debug.js';
 export * from './doc.js';
 export * from './enum.js';
 export * from './entity.js';
+export { default as prepareEntity } from './prepare-entity.js';
 export * from './entities.js';
 export * from './field-utils.js';
 export { default as prepareField } from './prepare-field.js';
